Rename _exit.c to _unistd_exit.c.
When I follow the commands (mentioned in https://github.com/SNSystems/hello-rld/tree/libc) to build hello-rld, I got an error of “Undefined symbol: _Exit”.

The problem is related with the command of "ar -x --output libc_repo libc_repo.a" command in the wsl2. During the command, the _exit.t overwrites the _Exit.t. It is caused by the fact that Windows (and macOS for thar matter) default file systems (NTFS and APFS) are case-preserving but not case-sensitive. That means that, on those systems, _Exit.t and _exit.t are the *same* file. Traditional Unix, and therefore Linux (ext4), file systems are case-sensitive so _Exit.t and exit.t are different files.

After renaming the _exit.c, the issue has been solved.