Allow an external REPOFILE variable to control the repository file.
Remove the exported database from the ALL_TICKET_LIBS
variable. There's no need to include a JSON file in our distribution
since this can be created simply using pstore-export.

Remove the directory name from the recipe for creating the
pre-assembled ticket files, preferring $(dir x) instad. This isn't too
important in most cases, but we shouldn't hard-wire the value of
$(JSON_CRT_TICKET) in two places.

Remove the rule that copies the database. Since the name and location
of the database can be controlled externally, this is no longer necessary.